Added (js_deployer)
This code is used to deploy openlibrary js code, by typing the name of
a regular openlibrary object. (Similar to a abstraction wrapper. It is
extremelly useful, for example, for o.pattrutils.js, which reads
patcher hierarchies. And if I used an abstraction wrapper I would have
to implement changes to js code, which would be inside of a patcher.)

- js_deployer.js
- o.library.js_deployer template
- documentation (/docs/about_jsobjects)
- also removed all objectmappings (except the file
"o-objectmappings.txt"), since I could not recreate the js_deployer
procedure through objectmappings